chore: order types
