+++ helpers/providers/anthropic.ts
@@ -1,7 +1,7 @@
 import ciInfo from "ci-info";
 import prompts from "prompts";
 import { ModelConfigParams } from ".";
-import { questionHandlers, toChoice } from "../../questions/questions";
+import { questionHandlers, toChoice } from "../../questions/utils";
 
 const MODELS = [
   "claude-3-opus",
